Cheat a bit on the dfs-datastores VersionedKeyValSourceTest
dfs-datastores' semantics have shifted since 1.3.4, and now it is of the consumer's responsibility to create
the empty version directory which will contain the version's data.
In this whitebox test, nobody was creating the empty directory anymore, so the test failed.

Added a "hint" and an explanation (not sure this tests much now, but keeps the original dance mostly in)